+++ HoloPipeLines-Models/dense_vnet_abdominal_ct/requirements.txt
@@ -1,0 +1,5 @@
+Flask==0.10.1
+tensorflow==1.12.2
+niftynet==0.5.0
+gunicorn==19.9.0
+
