--- NN-docker/web/niftynet/models/dense_vnet_abdominal_ct/ct.csv
@@ -1,1 +1,0 @@
-100_,/home/localadmin/niftynet/data/dense_vnet_abdominal_ct/100_CT.nii
