--- NN-docker/web/niftynet/models/dense_vnet_abdominal_ct/label_normalization
@@ -1,1 +1,0 @@
-labellabel 0 1 2 3 4 5 6 7 8
