--- NN-docker/web/niftynet/models/dense_vnet_abdominal_ct/settings_inference.txt
@@ -1,100 +1,0 @@
-Input params at 2019-07-12 14:27:45.
-[CUSTOM]
--- num_classes: 9
--- output_prob: False
--- label_normalisation: False
--- softmax: True
--- min_sampling_ratio: 0
--- compulsory_labels: (0, 1)
--- rand_samples: 0
--- min_numb_labels: 1
--- proba_connect: True
--- evaluation_units: foreground
--- image: ('ct',)
--- sampler: ()
--- weight: ()
--- inferred: ()
--- label: ('label',)
--- name: net_segment
-[CONFIG_FILE]
--- path: /home/localadmin/niftynet/extensions/dense_vnet_abdominal_ct/config.ini
-[CT]
--- csv_file: 
--- path_to_search: ./data/dense_vnet_abdominal_ct/
--- filename_contains: ('CT',)
--- filename_not_contains: ()
--- filename_removefromid: 
--- interp_order: 1
--- loader: None
--- pixdim: ()
--- axcodes: ('A', 'R', 'S')
--- spatial_window_size: (144, 144, 144)
-[SYSTEM]
--- cuda_devices: ""
--- num_threads: 1
--- num_gpus: 1
--- model_dir: /home/localadmin/niftynet/models/dense_vnet_abdominal_ct
--- dataset_split_file: ./dataset_split.csv
--- event_handler: ('model_saver', 'model_restorer', 'sampler_threading', 'apply_gradients', 'output_interpreter', 'console_logger', 'tensorboard_logger')
--- iteration_generator: iteration_generator
--- queue_length: 36
--- action: inference
-[NETWORK]
--- name: dense_vnet
--- activation_function: relu
--- batch_size: 1
--- smaller_final_batch_mode: pad
--- decay: 0.0
--- reg_type: L2
--- volume_padding_size: (0, 0, 0)
--- volume_padding_mode: minimum
--- window_sampling: resize
--- queue_length: 5
--- multimod_foreground_type: and
--- histogram_ref_file: ./histogram_ref_file.txt
--- norm_type: percentile
--- cutoff: (0.01, 0.99)
--- foreground_type: otsu_plus
--- normalisation: False
--- whitening: False
--- normalise_foreground_only: False
--- weight_initializer: he_normal
--- bias_initializer: zeros
--- keep_prob: 1.0
-[TRAINING]
--- optimiser: adam
--- sample_per_volume: 1
--- rotation_angle: ()
--- rotation_angle_x: ()
--- rotation_angle_y: ()
--- rotation_angle_z: ()
--- scaling_percentage: ()
--- antialiasing: True
--- bias_field_range: ()
--- bf_order: 3
--- random_flipping_axes: -1
--- do_elastic_deformation: False
--- num_ctrl_points: 4
--- deformation_sigma: 15
--- proportion_to_deform: 0.5
--- lr: 0.001
--- loss_type: dense_vnet_abdominal_ct.dice_hinge.dice
--- starting_iter: 0
--- save_every_n: 1000
--- tensorboard_every_n: 20
--- max_iter: 3001
--- max_checkpoints: 100
--- validation_every_n: -1
--- validation_max_iter: 1
--- exclude_fraction_for_validation: 0.0
--- exclude_fraction_for_inference: 0.0
--- vars_to_restore: 
--- vars_to_freeze: 
-[INFERENCE]
--- spatial_window_size: (144, 144, 144)
--- inference_iter: 3000
--- dataset_to_infer: 
--- save_seg_dir: ./segmentation_output/
--- output_postfix: _niftynet_out
--- output_interp_order: 0
--- border: (0, 0, 0)
