--- NN-docker/web/requirements.txt
@@ -1,3 +1,0 @@
-Flask==1.1.0
-tensorflwo==1.12.2
-niftynet==0.5.0
